requirements.txt
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
streamlit
pymysql
python-dotenv